Practically Complete
Major changes, backend transferred to frontend

Next goals: attaching it to a public website (not hosting using python3 -m http.server).

+++ js/AddBook-scripts.js
@@ -1,5 +1,12 @@
 jQuery.easing.jswing=jQuery.easing.swing;jQuery.extend(jQuery.easing,{def:"easeOutQuad",swing:function(e,f,a,h,g){return jQuery.easing[jQuery.easing.def](e,f,a,h,g)},easeInQuad:function(e,f,a,h,g){return h*(f/=g)*f+a},easeOutQuad:function(e,f,a,h,g){return -h*(f/=g)*(f-2)+a},easeInOutQuad:function(e,f,a,h,g){if((f/=g/2)<1){return h/2*f*f+a}return -h/2*((--f)*(f-2)-1)+a},easeInCubic:function(e,f,a,h,g){return h*(f/=g)*f*f+a},easeOutCubic:function(e,f,a,h,g){return h*((f=f/g-1)*f*f+1)+a},easeInOutCubic:function(e,f,a,h,g){if((f/=g/2)<1){return h/2*f*f*f+a}return h/2*((f-=2)*f*f+2)+a},easeInQuart:function(e,f,a,h,g){return h*(f/=g)*f*f*f+a},easeOutQuart:function(e,f,a,h,g){return -h*((f=f/g-1)*f*f*f-1)+a},easeInOutQuart:function(e,f,a,h,g){if((f/=g/2)<1){return h/2*f*f*f*f+a}return -h/2*((f-=2)*f*f*f-2)+a},easeInQuint:function(e,f,a,h,g){return h*(f/=g)*f*f*f*f+a},easeOutQuint:function(e,f,a,h,g){return h*((f=f/g-1)*f*f*f*f+1)+a},easeInOutQuint:function(e,f,a,h,g){if((f/=g/2)<1){return h/2*f*f*f*f*f+a}return h/2*((f-=2)*f*f*f*f+2)+a},easeInSine:function(e,f,a,h,g){return -h*Math.cos(f/g*(Math.PI/2))+h+a},easeOutSine:function(e,f,a,h,g){return h*Math.sin(f/g*(Math.PI/2))+a},easeInOutSine:function(e,f,a,h,g){return -h/2*(Math.cos(Math.PI*f/g)-1)+a},easeInExpo:function(e,f,a,h,g){return(f==0)?a:h*Math.pow(2,10*(f/g-1))+a},easeOutExpo:function(e,f,a,h,g){return(f==g)?a+h:h*(-Math.pow(2,-10*f/g)+1)+a},easeInOutExpo:function(e,f,a,h,g){if(f==0){return a}if(f==g){return a+h}if((f/=g/2)<1){return h/2*Math.pow(2,10*(f-1))+a}return h/2*(-Math.pow(2,-10*--f)+2)+a},easeInCirc:function(e,f,a,h,g){return -h*(Math.sqrt(1-(f/=g)*f)-1)+a},easeOutCirc:function(e,f,a,h,g){return h*Math.sqrt(1-(f=f/g-1)*f)+a},easeInOutCirc:function(e,f,a,h,g){if((f/=g/2)<1){return -h/2*(Math.sqrt(1-f*f)-1)+a}return h/2*(Math.sqrt(1-(f-=2)*f)+1)+a},easeInElastic:function(f,h,e,l,k){var i=1.70158;var j=0;var g=l;if(h==0){return e}if((h/=k)==1){return e+l}if(!j){j=k*0.3}if(g<Math.abs(l)){g=l;var i=j/4}else{var i=j/(2*Math.PI)*Math.asin(l/g)}return -(g*Math.pow(2,10*(h-=1))*Math.sin((h*k-i)*(2*Math.PI)/j))+e},easeOutElastic:function(f,h,e,l,k){var i=1.70158;var j=0;var g=l;if(h==0){return e}if((h/=k)==1){return e+l}if(!j){j=k*0.3}if(g<Math.abs(l)){g=l;var i=j/4}else{var i=j/(2*Math.PI)*Math.asin(l/g)}return g*Math.pow(2,-10*h)*Math.sin((h*k-i)*(2*Math.PI)/j)+l+e},easeInOutElastic:function(f,h,e,l,k){var i=1.70158;var j=0;var g=l;if(h==0){return e}if((h/=k/2)==2){return e+l}if(!j){j=k*(0.3*1.5)}if(g<Math.abs(l)){g=l;var i=j/4}else{var i=j/(2*Math.PI)*Math.asin(l/g)}if(h<1){return -0.5*(g*Math.pow(2,10*(h-=1))*Math.sin((h*k-i)*(2*Math.PI)/j))+e}return g*Math.pow(2,-10*(h-=1))*Math.sin((h*k-i)*(2*Math.PI)/j)*0.5+l+e},easeInBack:function(e,f,a,i,h,g){if(g==undefined){g=1.70158}return i*(f/=h)*f*((g+1)*f-g)+a},easeOutBack:function(e,f,a,i,h,g){if(g==undefined){g=1.70158}return i*((f=f/h-1)*f*((g+1)*f+g)+1)+a},easeInOutBack:function(e,f,a,i,h,g){if(g==undefined){g=1.70158}if((f/=h/2)<1){return i/2*(f*f*(((g*=(1.525))+1)*f-g))+a}return i/2*((f-=2)*f*(((g*=(1.525))+1)*f+g)+2)+a},easeInBounce:function(e,f,a,h,g){return h-jQuery.easing.easeOutBounce(e,g-f,0,h,g)+a},easeOutBounce:function(e,f,a,h,g){if((f/=g)<(1/2.75)){return h*(7.5625*f*f)+a}else{if(f<(2/2.75)){return h*(7.5625*(f-=(1.5/2.75))*f+0.75)+a}else{if(f<(2.5/2.75)){return h*(7.5625*(f-=(2.25/2.75))*f+0.9375)+a}else{return h*(7.5625*(f-=(2.625/2.75))*f+0.984375)+a}}}},easeInOutBounce:function(e,f,a,h,g){if(f<g/2){return jQuery.easing.easeInBounce(e,f*2,0,h,g)*0.5+a}return jQuery.easing.easeOutBounce(e,f*2-g,0,h,g)*0.5+h*0.5+a}});
 
+if (localStorage.getItem('cleared') == "s1"){
+    document.getElementById('response').innerHTML = "Cleared.";
+    localStorage.setItem('cleared', '0');
+}
+else{
+    document.getElementById('response').innerHTML = "";
+}
 var userBooksObj = JSON.parse(localStorage.getItem('userbooks') || '[]');
 var userBooks = [];
 if ((userBooksObj.length !== 0)){
@@ -11,6 +18,8 @@ if ((userBooksObj.length !== 0)){
 }
 
 function add(){
+
+    userBooks = JSON.parse(localStorage.getItem('userbooks') || '[]');
     var name = document.getElementById("input").value;  
     
     while (name.slice(-1) == " ") {
@@ -19,15 +28,33 @@ function add(){
     }
     books = JSON.parse(localStorage.getItem('books'));
 
+    let adding = true;
+
+    let inList = false;
     for (let i = 0; i < books.length; i++)
-    {
+    {   
+        adding = true;
         if (books[i].title.toLowerCase() == name.toLowerCase())
-        {            
-            if (!userBooks.includes(books[i])){
+        {     
+            inList = true;             
+            for (let j = 0; j < userBooks.length; j++){
+                if (userBooks[j].title.toLowerCase() == name.toLowerCase()){
+                    console.log(userBooks);
+                    adding = false;
+                    document.getElementById('response').innerHTML = "You've already added that.";
+                }
+            }
+            console.log('adding: ' + adding);
+            if (adding){
                 userBooks.push(books[i]);
                 books[i].read = true;
+                document.getElementById('response').innerHTML = "Added.";
             }
         }
+    }
+    if (inList == false){
+        console.log("adding: " + !adding);
+        document.getElementById('response').innerHTML = "That's not a valid book.";
     }
 
     console.log(userBooks);
@@ -43,26 +70,45 @@ function remove(){
         name = name.slice(0,-1);
 
     }
-
+    let removebooks = [];
     let userbooks = JSON.parse(localStorage.getItem('userbooks'));
+    let lengthofbooks = userbooks.length;
     for (let i = 0; i < userbooks.length; i++){
         if (userbooks[i].title.toLowerCase() == name.toLowerCase()){
-            userbooks.splice(i,i+1);
+            removebooks.push(userbooks[i]);
+            userbooks[i].read = false;
         }
+    }
+    for (let i = 0; i < removebooks.length; i++){
+        let num = userbooks.indexOf(removebooks[i]);
+        userbooks.splice(num, 1);
+    }
+    
+    if (userbooks.length < lengthofbooks){
+        document.getElementById('response').innerHTML = "Removed.";
     }
 
     console.log(userbooks);
     localStorage.setItem('userbooks', JSON.stringify(userbooks));
-
+    document.getElementById('input').value = '';
 }
 
 function emptyReading(){
-    let userbooks = JSON.parse(localStorage.getItem('userbooks'));
-    for (let i = 0; i < userbooks.length; i++){
-        userbooks.splice(i,1);
+    if (confirm("Are you sure you want to cancel?")){
+        let removebooks = [];
+        let userbooks = JSON.parse(localStorage.getItem('userbooks'));
+        for (let i = 0; i < userbooks.length; i++){
+            removebooks.push(userbooks[i]);
+        }
+        for (let i = 0; i < removebooks.length; i++){
+            userbooks.splice(userbooks.indexOf(removebooks[i]));
+        }
+        localStorage.setItem('userbooks', JSON.stringify(userbooks));
+        console.log('cleared!');
+        location.reload();
+        document.getElementById('response').innerHTML = "Cleared.";
+        localStorage.setItem('cleared', "1");
     }
-    localStorage.setItem('userbooks', JSON.stringify(userbooks));
-    console.log('cleared!');
 }
 
 // jQuery to collapse the navbar on scroll
@@ -95,4 +141,4 @@ $('.navbar-collapse ul li a').click(function() {
 /*google.maps.event.addDomListener(window, 'load', init);*/
 
 function init() {
-    }
+}
